Backend/requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.2
blinker==1.6.2
blis==0.7.11
cachelib==0.10.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.3
cymem==2.0.8
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
eventregistry==9.1
filelock==3.12.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-Session==0.6.0
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.21.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
loguru==0.7.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openai==0.28.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
peft==0.6.2
Pillow==10.0.1
pinecone-client==2.2.2
pinecone-io==0.0.1
preshed==3.0.9
psutil==5.9.6
pydantic==2.4.1
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.38.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wasabi==1.1.2
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
gunicorn
xformers